Refactor: move weather logic to the new `useWeather` hook
